:bug: handles np float types when serializing JSON

+++ tests/annotator_tests/openpose_tests/json_encode_test.py
@@ -1,5 +1,6 @@
 import json
 import unittest
+import numpy as np
 
 import importlib
 utils = importlib.import_module('extensions.sd-webui-controlnet.tests.utils', 'utils')
@@ -41,7 +42,7 @@ class TestEncodePosesAsJson(unittest.TestCase):
         self.assertEqual(result, expected)
 
     def test_single_pose_with_keypoints(self):
-        keypoints = [Keypoint(0.5, 0.5), None, Keypoint(0.6, 0.6)]
+        keypoints = [Keypoint(np.float32(0.5), np.float32(0.5)), None, Keypoint(0.6, 0.6)]
         poses = [PoseResult(BodyResult(keypoints, 0, 0), keypoints, keypoints, keypoints)]
         canvas_height = 1080
         canvas_width = 1920
